Rename slugs.ts to convert-slug-terms.ts for clarity
